+++ src/app/selector_win.py
@@ -555,7 +555,7 @@ class selWin:
         self.wid_scroll.grid(row=0, column=1, sticky="NS")
         self.wid_canvas['yscrollcommand'] = self.wid_scroll.set
 
-        utils.add_mousewheel(self.wid_canvas, self.win)
+        tk_tools.add_mousewheel(self.wid_canvas, self.win)
 
         if utils.MAC:
             # Labelframe doesn't look good here on OSX
